backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
anthropic==0.34.2
python-dotenv==1.0.1
pytest==8.3.3
pytest-asyncio==0.24.0
